TIP-701: move pqb factory into elasticsearch folder
